Change data display method of TableView; Drop classes TableTab and TableModel
In TableView, data will now be displayed via QSqlTableModel and a read-only
variant of QSqlRelationalTableModel. This leads to a more flexible way of
storing data. Furthermore the Views Buecher, Lehrer, Schueler, LAusleihe, and
SAusleihe can be dropped now. Nevertheless, the new system needs the additional
Views Buch (with concatenated title and form), Sausleihe, and Btausch (both with
two `sid` fields.
Moreover, the Schülerausleihe and Büchertausch tabs will now also display the
title of the book and its form.