Add buecher.sql; Change behaviour of ListModel
The data of a ListModel instance will not be saved in a list of integers, but
in an associative array of associative arrays of booleans, which means that it
is now possible for a form to have more than 31 books (limit of int).
Furthermore this way is less vulnerable and the view BTausch can now be dropped.