Fix export functionality
I changed the export method so that the correct form is
displayed (and not the address of the combo box).
Furthermore, I added projection to the SQL queries
because appearently QSqlQuery doesn't handle multiple
tables together with SELECT * ... so well. Now, only the
names of the students and the books' titles will be
fetched.